requirements.txt
matplotlib
seaborn
numpy
pandas
streamlit
ultralytics
ipykernel
Pillow